add auth module
