+++ src/app.module.ts
@@ -4,7 +4,9 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 
 import configuration from './config/configuration';
 import { UsersModule } from './users/users.module';
+import { AuthModule } from './auth/auth.module';
 import { User } from './users/entities/user.entity';
+import { JwtAccessTokenStrategy } from './auth/strategies/jwt-access.strategy';
 
 @Module({
   imports: [
@@ -26,7 +28,9 @@ import { User } from './users/entities/user.entity';
       }),
       inject: [ConfigService],
     }),
+    AuthModule,
     UsersModule,
   ],
+  providers: [JwtAccessTokenStrategy],
 })
 export class AppModule {}
