+++ src/auth/auth.module.ts
@@ -1,0 +1,25 @@
+import { Module } from '@nestjs/common';
+import { AuthController } from './auth.controller';
+import { AuthService } from './auth.service';
+import { JwtRefreshTokenStrategy } from './strategies/jwt-refresh.strategy';
+import { GoogleOauthStrategy } from './strategies/google-oauth.strategy';
+import { UsersModule } from '../users/users.module';
+import { JwtModule } from '@nestjs/jwt';
+import { ConfigModule, ConfigService } from '@nestjs/config';
+
+@Module({
+  imports: [
+    JwtModule.registerAsync({
+      imports: [ConfigModule],
+      useFactory: (configService: ConfigService) => ({
+        secret: configService.get('jwtSecret'),
+        signOptions: { expiresIn: '1h' },
+      }),
+      inject: [ConfigService],
+    }),
+    UsersModule,
+  ],
+  controllers: [AuthController],
+  providers: [AuthService, JwtRefreshTokenStrategy, GoogleOauthStrategy],
+})
+export class AuthModule {}
