+++ src/users/decorators/user-param.decorator.ts
@@ -1,3 +1,3 @@
-import { createRequestParamDecorator } from 'src/decorators/create-request-param.decorator';
+import { createRequestParamDecorator } from 'src/common/decorators/create-request-param.decorator';
 
 export const User = createRequestParamDecorator('user');
