+++ src/users/interceptors/serialize-user.interceptor.ts
@@ -1,6 +1,6 @@
 import { UseInterceptors } from '@nestjs/common';
 
-import { SerializeInterceptor } from '../../interceptors/serialize.intercetor';
+import { SerializeInterceptor } from '../../common/interceptors/serialize.intercetor';
 import { UserDto } from '../dtos/user.dto';
 
 export function SerializeUser() {
