+++ src/users/users.controller.ts
@@ -1,7 +1,6 @@
 import {
   Controller,
   Get,
-  Post,
   Put,
   Delete,
   Body,
@@ -9,41 +8,18 @@ import {
   UseFilters,
 } from '@nestjs/common';
 
+import { HttpExceptionFilter } from '../common/filters/http-exception.filter';
 import { UpdateUserDto } from './dtos/update-user.dto';
 import { TokenUserDto } from './dtos/token-user.dto';
-import { LoginUserDto } from './dtos/login-user.dto';
 import { UsersService } from './users.service';
-import { AuthService } from './auth.service';
-import { JwtAccessTokenAuthGuard } from './guards/jwt-access-auth.guard';
-import { JwtRefreshTokenAuthGuard } from './guards/jwt-refresh-auth.guard';
+import { JwtAccessTokenAuthGuard } from '../auth/guards/jwt-access-auth.guard';
 import { SerializeUser } from './interceptors/serialize-user.interceptor';
-import { HttpExceptionFilter } from '../filters/http-exception.filter';
 import { User } from './decorators/user-param.decorator';
-import { GoogleOAuthGuard } from './guards/google-oauth.guard';
 
 @Controller('users')
 @UseFilters(new HttpExceptionFilter())
 export class UsersController {
-  constructor(
-    private usersService: UsersService,
-    private authService: AuthService,
-  ) {}
-
-  @Get('google')
-  @UseGuards(GoogleOAuthGuard)
-  async login() {}
-
-  @Get('google/callback')
-  @UseGuards(GoogleOAuthGuard)
-  async loginCallback(@User() user: LoginUserDto) {
-    return this.authService.login(user);
-  }
-
-  @Post('refresh')
-  @UseGuards(JwtRefreshTokenAuthGuard)
-  async refresh(@User() user: TokenUserDto) {
-    return this.authService.getAccessToken(user);
-  }
+  constructor(private usersService: UsersService) {}
 
   @Get('self')
   @UseGuards(JwtAccessTokenAuthGuard)
