+++ src/users/users.module.ts
@@ -1,35 +1,14 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import { JwtModule } from '@nestjs/jwt';
 
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
-import { AuthService } from './auth.service';
 import { User } from './entities/user.entity';
-import { JwtAccessTokenStrategy } from './strategies/jwt-access.strategy';
-import { JwtRefreshTokenStrategy } from './strategies/jwt-refresh.strategy';
-import { GoogleOauthStrategy } from './strategies/google-oauth.strategy';
 
 @Module({
-  imports: [
-    TypeOrmModule.forFeature([User]),
-    JwtModule.registerAsync({
-      imports: [ConfigModule],
-      useFactory: (configService: ConfigService) => ({
-        secret: configService.get('jwtSecret'),
-        signOptions: { expiresIn: '1h' },
-      }),
-      inject: [ConfigService],
-    }),
-  ],
+  imports: [TypeOrmModule.forFeature([User])],
   controllers: [UsersController],
-  providers: [
-    UsersService,
-    AuthService,
-    JwtAccessTokenStrategy,
-    JwtRefreshTokenStrategy,
-    GoogleOauthStrategy,
-  ],
+  providers: [UsersService],
+  exports: [UsersService],
 })
 export class UsersModule {}
